+++ test/unit/test_unit.ml
@@ -110,11 +110,154 @@ module Test_noit = struct
   let tests = test_dump @ test_roots @ test_children
 end
 
+let reindent ~source ~range:(low, high) indents =
+  let lines = String.split_lines source in
+  let low = low - 1 and high = high - 1 in
+  let lines =
+    List.mapi lines ~f:(fun i line ->
+        if i < low then line
+        else if low <= i && i <= high then
+          let indent = List.nth_exn indents (i - low) in
+          let line = String.lstrip line in
+          let spaces = String.make indent ' ' in
+          spaces ^ line
+        else line )
+  in
+  String.concat ~sep:"\n" lines
+
+module Test_translation_unit = struct
+  let test_indent =
+    let test name ~source ~range expected =
+      let test_name = "indent: " ^ name in
+      ( test_name
+      , `Quick
+      , fun () ->
+          let got =
+            Translation_unit.indentation Use_file ~input_name:"_" ~source
+              ~range Conf.ocamlformat_profile Conf.default_opts
+            |> Result.map ~f:(reindent ~source ~range)
+          in
+          Alcotest.check
+            Alcotest.(result string Testable.error)
+            test_name expected got )
+    in
+    [ test "invalid low" ~source:"foo\nbar" ~range:(0, 1)
+        (Error (User_error "Invalid line number 0."))
+    ; test "invalid high" ~source:"foo\nbar" ~range:(1, 4)
+        (Error (User_error "Invalid line number 4."))
+    ; test "invalid range" ~source:"foo\nbar" ~range:(2, 1)
+        (Error (User_error "Invalid range 2-1."))
+    ; test "empty buffer" ~source:"" ~range:(1, 1) (Ok "")
+    ; test "last buffer line" ~source:"foo\nbar" ~range:(2, 3) (Ok "foo\nbar")
+    ; test "already formatted"
+        ~source:
+          {|let foooooo =
+  let baaaaar =
+    let woooooo = foooooo in
+    let xooooo = bar + foo in
+    woooooo
+  in
+  bar
+|}
+        ~range:(1, 7)
+        (Ok
+           {|let foooooo =
+  let baaaaar =
+    let woooooo = foooooo in
+    let xooooo = bar + foo in
+    woooooo
+  in
+  bar|}
+        )
+    ; test "not already formatted"
+        ~source:
+          {|let foooooooooooo = let foooooooo = foooooooooooo in foooooooooooo
+
+let foooooooooooo = let foooooooooooooo =
+let woooooooooooooooo = koooooooooooooooo in
+baaaaaar
+in
+hooohoooo
+|}
+        ~range:(1, 7)
+        (Ok
+           {|let foooooooooooo = let foooooooo = foooooooooooo in foooooooooooo
+
+let foooooooooooo = let foooooooooooooo =
+    let woooooooooooooooo = koooooooooooooooo in
+    baaaaaar
+  in
+  hooohoooo|}
+        )
+    ; test "with parens and begin/end"
+        ~source:
+          {|let x = begin
+    let y =
+      (if (k = x)
+       then
+         begin match k,v with [x; y] -> ( foo;
+                                        (if (z) then foo else bar) )
+         end
+       else
+         foo)
+    in
+    foooooo
+  end|}
+        ~range:(1, 12)
+        (Ok
+           {|let x = begin
+  let y =
+    (if (k = x)
+    then
+      begin match k,v with [x; y] -> ( foo;
+          (if (z) then foo else bar) )
+    end
+    else
+      foo)
+  in
+  foooooo
+  end|}
+        )
+    ; test "split over multiple lines"
+        ~source:{|let fooooo =
+[
+foooooo ;
+foooooooo ;
+fooooooo
+]|}
+        ~range:(1, 6)
+        (Ok {|let fooooo =
+  [
+    foooooo ;
+    foooooooo ;
+    fooooooo
+]|})
+    ; test "invalid file"
+        ~source:{|let foooooo =
+let foooooooooooo =
+(
+[
+fun x ->
+foooooo|}
+        ~range:(1, 6)
+        (Ok
+           {|let foooooo =
+  let foooooooooooo =
+    (
+      [
+        fun x ->
+          foooooo|}
+        ) ]
+
+  let tests = test_indent
+end
+
 let tests =
   [ ("Location", Test_location.tests)
   ; ("non overlapping interval tree", Test_noit.tests)
   ; ("Ast", Test_ast.tests)
   ; ("Literal_lexer", Test_literal_lexer.tests)
-  ; ("Fmt", Test_fmt.tests) ]
+  ; ("Fmt", Test_fmt.tests)
+  ; ("Translation_unit", Test_translation_unit.tests) ]
 
 let () = Alcotest.run "ocamlformat" tests