+++ test/unit/testable.mli
@@ -1,0 +1,1 @@
+val error : Ocamlformat_lib.Translation_unit.error Alcotest.testable