+++ CHANGES.md
@@ -41,6 +41,10 @@
 
   + Consistent positioning of lambda return type annotations when no-break-infix-before-func and pre/post extensions (#1581, @gpetiot)
 
+#### New features
+
+  + Line/region indentation feature (emacs integration) (#1207, @gpetiot)
+
 ### 0.16.0 (2020-11-16)
 
 #### Removed